+++ Source/hydro/advection_util_nd.F90
@@ -16,7 +16,8 @@ contains
 #ifndef AMREX_USE_GPU
     use castro_error_module, only: castro_error
 #endif
-    use amrex_fort_module, only: rt => amrex_real, amrex_min
+    use amrex_fort_module, only: rt => amrex_real
+    use reduction_module, only: reduce_min
 
     implicit none
 
@@ -159,7 +160,7 @@ contains
                 ! Store the maximum (negative) fractional change in the density from this reset.
 
                 if (old_rho < ZERO) then
-                   call amrex_min(frac_change, (state(i,j,k,URHO) - old_rho) / old_rho)
+                   call reduce_min(frac_change, (state(i,j,k,URHO) - old_rho) / old_rho)
                 end if
 
              end if
@@ -296,7 +297,8 @@ contains
     use amrex_constants_module, only: ZERO, ONE
     use meth_params_module, only: NQ, QRHO, QU, QV, QW, QC, NQAUX, time_integration_method
     use prob_params_module, only: dim
-    use amrex_fort_module, only : rt => amrex_real, amrex_max
+    use amrex_fort_module, only : rt => amrex_real
+    use reduction_module, only: reduce_max
 
     implicit none
 
@@ -410,14 +412,14 @@ contains
                 end if
 #endif
 
-                call amrex_max(courno, courtmp)
+                call reduce_max(courno, courtmp)
              endif
           enddo
        enddo
     enddo
 
     if (time_integration_method == 0) then
-       call amrex_max(courno, max(courmx, courmy, courmz))
+       call reduce_max(courno, max(courmx, courmy, courmz))
     endif
 
   end subroutine ca_compute_cfl