+++ CHANGES.md
@@ -1,5 +1,10 @@
 # changes since the last release
 
+   * A subroutine eos_on_host has been added to the EOS module.
+     This is a wrapper for the EOS that must be used for CUDA
+     builds if the EOS is being called in probinit or other
+     places that don't run on the GPU. (#693)
+
    * We now use VODE90 instead of VODE by default. (#677)
 
 # 19.11